app/requirements.txt

boto3==1.28.72
botocore==1.31.72

fastapi==0.104.0
gunicorn==21.2.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
mlflow==2.15.1
numpy==1.26.0
pandas==2.1.1
Pillow==10.1.0
pip==23.3
psycopg2-binary==2.9.7
pydantic==2.4.2
pydantic_core==2.10.1
scikit-learn==1.3.1
scipy==1.11.3
streamlit==1.32.0
tomli==2.0.1
uvicorn==0.23.2
pendulum
